"SPI communication for DAC: added ClearDAC function to set CLR pin low and write zero scale to all input and DAC registers"
This commit adds a new function, ClearDAC(), which sets the CLR (clear) pin low to write zero scale to all input and DAC registers. This is useful for resetting the DAC to a known state. The function is called in the main program. Additionally, the CLR pin is set high in the Send2DAC() function to enable DAC registers.